Cleanup install script
This is a work in progress that prefers to use `.config` over the home
directory. Currently only the git and i3 configurations are linked.

The creation of directories also changed from a single source directory
per language to a single source directory for all languages. The
language used shouldn't really matter.