Only add to $PATH if the directory is not already in it
For this purpose I've added the __add_to_path function, which first
checks if a path is present in $PATH and only adds it if it's not
present already.

Before this commit we blindly added directories to $PATH which is a
problem when source profile twice, which always happens: first from the
terminal emulator (and thus bash) and then when starting tmux (and thus
bash again).